pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "depth2normal"
version = "1.0.0"
description = "Convert depth map images to normal map images"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Mert Cobanov" }]
keywords = ["depth-map", "normal-map", "3d", "computer-vision", "image-processing"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Typing :: Typed",
]
dependencies = [
    "numpy",
    "Pillow",
    "scipy",
    "click",
]

[project.scripts]
depth2normal = "depth2normal.cli:cli"

[project.urls]
Homepage = "https://github.com/cobanov/depth2normal"
Repository = "https://github.com/cobanov/depth2normal"
Issues = "https://github.com/cobanov/depth2normal/issues"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]
